+++ customRandomContraints.py
@@ -1,0 +1,3 @@
+
+def doPostProcessing(typeName, obj):
+    return True, obj